+++ CHANGELOG.md
@@ -92,3 +92,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added secure API to retrieve bearer token.
+- Create secure API to retrieve list of emails associated with a form.